Refactor mobile UI album action sheets into module
